+++ app/signup/page.tsx
@@ -1,48 +1,15 @@
-import { createClient } from '@/utils/supabase/server'
+'use client'
+
 import Link from 'next/link'
 import { SubmitButton } from '../../components/forms/submit-button'
 import { Input } from '@/components/forms/input'
 import { Label } from '@/components/forms/label'
-import { FormMessage, Message } from '@/components/forms/form-message'
-import { encodedRedirect } from '@/utils/utils'
-import { cookies } from 'next/headers'
+import { useFormState } from 'react-dom'
+import signUp from '@/app/actions'
 
-export default function Signup({ searchParams }: { searchParams: Message }) {
-  const signUp = async (formData: FormData) => {
-    'use server'
-    const phone = formData.get('phone')?.toString()
-    const supabase = createClient()
-
-    if (!phone) {
-      return { error: 'Phone are required' }
-    }
-
-    // https://supabase.com/docs/guides/auth/phone-login#signing-in-with-phone-otp
-    const { error } = await supabase.auth.signInWithOtp({
-      phone,
-    })
-
-    if (error) {
-      console.error(error.code + ' ' + error.message)
-      return encodedRedirect('error', '/signup', 'Error trying to sign up')
-    } else {
-      cookies().set('phone', phone)
-      cookies().set('verification_process', 'sms')
-      return encodedRedirect(
-        'success',
-        '/verify',
-        'Thanks for signing is! Please check your SMS for a verification code.'
-      )
-    }
-  }
-
-  if ('message' in searchParams) {
-    return (
-      <div className="w-full flex-1 flex items-center h-screen sm:max-w-md justify-center gap-2 p-4">
-        <FormMessage message={searchParams} />
-      </div>
-    )
-  }
+export default function Signup() {
+  const initialState = { error: '' }
+  const [signUpState, signUpDispatch] = useFormState(signUp, initialState)
 
   return (
     <div className="w-full flex-1 flex items-center h-screen sm:max-w-md justify-center gap-2 p-4">
@@ -67,16 +34,23 @@ export default function Signup({ searchParams }: { searchParams: Message }) {
         Back
       </Link>
 
-      <form className="flex flex-col w-full justify-center gap-2 text-foreground [&>input]:mb-6 max-w-md">
+      <form
+        className="flex flex-col w-full justify-center gap-2 text-foreground [&>input]:mb-6 max-w-md"
+        action={signUpDispatch}
+      >
         <h1 className="text-2xl font-medium">Sign up/Log in</h1>
         <div className="mt-8 flex flex-col gap-2 [&>input]:mb-3">
           <Label htmlFor="phone">Phone</Label>
           <Input name="phone" placeholder="+810812345678" required />
-          <SubmitButton formAction={signUp} pendingText="Signing up...">
-            Sign up
+          <SubmitButton pendingText="Submitting...">
+            Sign up/Log in
           </SubmitButton>
         </div>
-        <FormMessage message={searchParams} />
+        {'error' in signUpState && signUpState.error !== '' && (
+          <div className="text-red-500 border-l-2 border-red-500 px-4">
+            {signUpState.error}
+          </div>
+        )}
       </form>
     </div>
   )
